Data-Modeling-with-DBT-Handling-440K-Records-main/requirements.txt
psycopg2-binary
sqlalchemy
pandas
PyPI
python-dotenv
dbt-core
dbt-postgres